requirements.txt
discord.py[speed]==2.0.1
python-dotenv
colorlog
uvloop==0.17.0; sys_platform != "win32" and platform_python_implementation == "CPython"
